generate create command

+++ src/main/java/com/kenshoo/pl/intellij/view/NewEntityForm.java
@@ -45,7 +45,7 @@ public class NewEntityForm {
         fieldTypes.addItem(new FieldType("INTEGER", Integer.class));
         fieldTypes.addItem(new FieldType("VARCHAR", String.class));
         fieldTypes.addItem(new FieldType("BOOLEAN", Boolean.class));
-        fieldTypes.addItem(new FieldType("FLOAT", Float.class));
+        fieldTypes.addItem(new FieldType("FLOAT", Double.class));
         fieldTypes.addItem(new FieldType("BIGINT", Long.class));
         fieldTypes.addItem(new FieldType("TIMESTAMP", Timestamp.class));
         return fieldTypes;
